# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/offers": {
      "get": {
        "operationId": "OffersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Rows per page, max 100.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "example": "games",
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "schema": {
              "example": "android",
              "type": "string",
              "enum": [
                "ios",
                "android",
                "web"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "coin",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OfferDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List live offers available to me",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/offers/categories": {
      "get": {
        "operationId": "OffersController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the categories that currently have offers",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/offers/mine": {
      "get": {
        "operationId": "OffersController_mine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Rows per page, max 100.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MyOfferDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Offers I started, with their reward status",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/offers/{id}": {
      "get": {
        "operationId": "OffersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OfferDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one offer",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/offers/{id}/click": {
      "post": {
        "operationId": "OffersController_click",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OfferClickDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start an offer and get its tracking URL",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/offers/postback/{provider_id}": {
      "get": {
        "operationId": "OffersController_postbackGet",
        "parameters": [
          {
            "name": "provider_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Provider reward webhook (GET). Public.",
        "tags": [
          "offers"
        ]
      },
      "post": {
        "operationId": "OffersController_postbackPost",
        "parameters": [
          {
            "name": "provider_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Provider reward webhook (POST). Public.",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/daily/challenges": {
      "get": {
        "operationId": "DailyController_challengeBoard",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "A past UTC day, yyyy-MM-dd. Defaults to today.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChallengeBoardDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The whole Daily Challenge screen — tiles, progress, master chest, calendar and what is playable",
        "tags": [
          "daily"
        ]
      }
    },
    "/api/daily/challenges/chest": {
      "post": {
        "operationId": "DailyController_claimChest",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChestClaimDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim the daily master chest, once every challenge is finished",
        "tags": [
          "daily"
        ]
      }
    },
    "/api/daily/challenges/{id}/claim": {
      "post": {
        "operationId": "DailyController_claimChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim the reward for a completed challenge",
        "tags": [
          "daily"
        ]
      }
    },
    "/api/daily/streak": {
      "get": {
        "operationId": "DailyController_streakBoard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/StreakBoardDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My 30-day streak board — the whole Rewards screen, called on app open",
        "tags": [
          "daily"
        ]
      }
    },
    "/api/daily/streak/claim": {
      "post": {
        "operationId": "DailyController_claimStreak",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/StreakClaimDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim today's streak day. This is the daily check-in.",
        "tags": [
          "daily"
        ]
      }
    },
...
    "/api/admin/daily/challenges": {
      "get": {
        "operationId": "AdminDailyController_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The daily challenge tiles, the master chest, and recent activity (admin only)",
        "tags": [
          "admin-daily"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coinzu API",
    "description": "Coinzu backend API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OfferGoalDto": {
        "type": "object",
        "properties": {
          "goal_id": {
            "type": "string",
            "example": "g1"
          },
          "title": {
            "type": "string",
            "example": "Install the app"
          },
          "reward_coins": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "goal_id",
          "title",
          "reward_coins"
        ]
      },
      "OfferDto": {
        "type": "object",
        "properties": {
          "cz_offer_id": {
            "type": "string",
            "format": "uuid"
          },
          "provider_id": {
            "type": "string",
            "format": "uuid"
          },
          "external_offer_id": {
            "type": "string",
            "example": "RT-99120"
          },
          "title": {
            "type": "string",
            "example": "Reach level 10 in Coin Quest"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "reward_coins": {
            "type": "number",
            "example": 2500
          },
          "reward_gems": {
            "type": "number",
            "example": 0
          },
          "category": {
            "type": "string",
            "nullable": true,
            "example": "games"
          },
          "countries": {
            "example": [
              "US"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "platforms": {
            "example": [
              "android"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "goals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfferGoalDto"
            }
          },
          "is_active": {
            "type": "boolean"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "cz_offer_id",
          "provider_id",
          "external_offer_id",
          "title",
          "description",
          "image_url",
          "reward_coins",
          "reward_gems",
          "category",
          "countries",
          "platforms",
          "goals",
          "is_active",
          "expires_at",
          "synced_at"
        ]
      },
      "MyOfferDto": {
        "type": "object",
        "properties": {
          "cz_offer_click_id": {
            "type": "string",
            "format": "uuid"
          },
          "clicked_at": {
            "type": "string",
            "format": "date-time"
          },
          "offer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OfferDto"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "pending",
              "approved",
              "reversed"
            ]
          },
          "payout_coins": {
            "type": "number",
            "example": 0
          },
          "credited_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "cz_offer_click_id",
          "clicked_at",
          "offer",
          "status",
          "payout_coins",
          "credited_at"
        ]
      },
      "OfferClickDto": {
        "type": "object",
        "properties": {
          "click_id": {
            "type": "string",
            "example": "cz_9f2c8a11c4d7b3e5"
          },
          "tracking_url": {
            "type": "string",
            "example": "https://track.rewardtym.com/click?offer=RT-99120"
          }
        },
        "required": [
          "click_id",
          "tracking_url"
        ]
      },
      "MasterChestDto": {
        "type": "object",
        "properties": {
          "reward_coins": {
            "type": "number",
            "example": 1000
          },
          "reward_gems": {
            "type": "number",
            "example": 1000
          },
          "completed": {
            "type": "number",
            "example": 3
          },
          "total": {
            "type": "number",
            "example": 5
          },
          "progress_label": {
            "type": "string",
            "example": "3/5"
          },
          "percent": {
            "type": "number",
            "example": 60
          },
          "is_unlocked": {
            "type": "boolean"
          },
          "is_claimed": {
            "type": "boolean"
          },
          "can_claim": {
            "type": "boolean"
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "reward_coins",
          "reward_gems",
          "completed",
          "total",
          "progress_label",
          "percent",
          "is_unlocked",
          "is_claimed",
          "can_claim",
          "claimed_at"
        ]
      },
      "MaxEarningDto": {
        "type": "object",
        "properties": {
          "coins": {
            "type": "number",
            "example": 3500
          },
          "gems": {
            "type": "number",
            "example": 1500
          }
        },
        "required": [
          "coins",
          "gems"
        ]
      },
      "ChallengeAvailabilityDto": {
        "type": "object",
        "properties": {
          "spin_available": {
            "type": "boolean"
          },
          "spins_left": {
            "type": "number",
            "example": 1
          },
          "quiz_available": {
            "type": "boolean"
          },
          "quiz_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scratch_cards_left": {
            "type": "number",
            "example": 0
          },
          "scratch_cards_pending": {
            "type": "number",
            "example": 0
          },
          "scratch_available": {
            "type": "boolean"
          }
        },
        "required": [
          "spin_available",
          "spins_left",
          "quiz_available",
          "quiz_id",
          "scratch_cards_left",
          "scratch_cards_pending",
          "scratch_available"
        ]
      },
      "ChallengeDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-28"
          },
          "label": {
            "type": "string",
            "example": "28"
          },
          "completed": {
            "type": "number",
            "example": 3
          },
          "total": {
            "type": "number",
            "example": 5
          },
          "is_today": {
            "type": "boolean"
          },
          "is_locked": {
            "type": "boolean"
          },
          "is_viewable": {
            "type": "boolean"
          }
        },
        "required": [
          "date",
          "label",
          "completed",
          "total",
          "is_today",
          "is_locked",
          "is_viewable"
        ]
      },
...
      "ChallengeBoardDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-28"
          },
          "is_today": {
            "type": "boolean"
          },
          "is_read_only": {
            "type": "boolean"
          },
          "streak_days": {
            "type": "number",
            "example": 11
          },
          "master_chest": {
            "$ref": "#/components/schemas/MasterChestDto"
          },
          "max_earning": {
            "$ref": "#/components/schemas/MaxEarningDto"
          },
          "availability": {
            "$ref": "#/components/schemas/ChallengeAvailabilityDto"
          },
          "calendar": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeDayDto"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeTileDto"
            }
          },
          "total": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "date",
          "is_today",
          "is_read_only",
          "streak_days",
          "master_chest",
          "max_earning",
          "availability",
          "calendar",
          "data",
          "total"
        ]
      },
      "ChestClaimDto": {
        "type": "object",
        "properties": {
          "claimed": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "example": "2026-09-28"
          },
          "reward_coins": {
            "type": "number",
            "example": 1000
          },
          "reward_gems": {
            "type": "number",
            "example": 1000
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "claimed",
          "date",
          "reward_coins",
          "reward_gems",
          "claimed_at"
        ]
      },
      "StreakDayDto": {
        "type": "object",
        "properties": {
          "day_number": {
            "type": "number",
            "example": 1
          },
          "reward_coins": {
            "type": "number",
            "example": 100
          },
          "reward_gems": {
            "type": "number",
            "example": 0
          },
          "label": {
            "type": "string",
            "nullable": true,
            "example": "Week 1 bonus"
          },
          "is_milestone": {
            "type": "boolean"
          },
          "is_claimed": {
            "type": "boolean"
          },
          "is_today": {
            "type": "boolean"
          }
        },
        "required": [
          "day_number",
          "reward_coins",
          "reward_gems",
          "label",
          "is_milestone",
          "is_claimed",
          "is_today"
        ]
      },
      "StreakBoardDto": {
        "type": "object",
        "properties": {
          "current_day": {
            "type": "number",
            "example": 11
          },
          "longest_day": {
            "type": "number",
            "example": 14
          },
          "last_claimed_date": {
            "type": "string",
            "nullable": true,
            "example": "2026-09-27"
          },
          "can_claim_today": {
            "type": "boolean"
          },
          "claimable_day": {
            "type": "number",
            "nullable": true,
            "example": 12
          },
          "cycle_days": {
            "type": "number",
            "example": 30
          },
          "total_coins": {
            "type": "number",
            "example": 5000
          },
          "total_gems": {
            "type": "number",
            "example": 2000
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StreakDayDto"
            }
          },
          "total": {
            "type": "number",
            "example": 30
          }
        },
        "required": [
          "current_day",
          "longest_day",
          "last_claimed_date",
          "can_claim_today",
          "claimable_day",
          "cycle_days",
          "total_coins",
          "total_gems",
          "data",
          "total"
        ]
      },
      "StreakClaimDto": {
        "type": "object",
        "properties": {
          "day_number": {
            "type": "number",
            "example": 12
          },
          "reward_coins": {
            "type": "number",
            "example": 100
          },
          "reward_gems": {
            "type": "number",
            "example": 0
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "is_milestone": {
            "type": "boolean"
          },
          "current_day": {
            "type": "number",
            "example": 12
          },
          "longest_day": {
            "type": "number",
            "example": 14
          },
          "cycle_days": {
            "type": "number",
            "example": 30
          },
          "cycle_completed": {
            "type": "boolean"
          },
          "claimed_date": {
            "type": "string",
            "example": "2026-09-28"
          },
          "total_checkins": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "day_number",
          "reward_coins",
          "reward_gems",
          "label",
          "is_milestone",
          "current_day",
          "longest_day",
          "cycle_days",
          "cycle_completed",
          "claimed_date",
          "total_checkins"
        ]
      },
...
    }
  }
}